# src/main/resources/openapi.yml
openapi: 3.0.3
info:
  title: API Gestión de Tareas - NUEVO SPA
  description: API RESTful para gestión de tareas con autenticación JWT
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /auth/login:
    post:
      tags:
        - Autenticación
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Token generado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'

  /tareas:
    get:
      tags:
        - Tareas
      summary: Listar tareas del usuario autenticado
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de tareas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TareaResponse'

    post:
      tags:
        - Tareas
      summary: Crear una nueva tarea
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TareaRequest'
      responses:
        '200':
          description: Tarea creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TareaResponse'

  /tareas/{id}:
    put:
      tags:
        - Tareas
      summary: Actualizar una tarea
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TareaRequest'
      responses:
        '200':
          description: Tarea actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TareaResponse'

    delete:
      tags:
        - Tareas
      summary: Eliminar una tarea
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Tarea eliminada correctamente

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    LoginResponse:
      type: object
      properties:
        token:
          type: string

    TareaRequest:
      type: object
      properties:
        titulo:
          type: string
        descripcion:
          type: string
        estadoId:
          type: integer
      required:
        - titulo
        - estadoId

    TareaResponse:
      type: object
      properties:
        id:
          type: integer
        titulo:
          type: string
        descripcion:
          type: string
        estado:
          type: string